Use Rollup for UMD production
We're about to introduce some additional modules to the Crushinator
helper, so we need a build system that converts ES6 modules to UMD.

Until now I've been using the "transform-es2015-modules-umd" plugin,
which simply loops through each of the imported ES6 modules and
wraps them in a UMD shim to be concatenated together. This is fine
for single-file JS libraries and for some other project types, but
in this case it would have meant our additional modules would each,
individually, have been leaked to the global / window object where
neither AMD nor CommonJS are present, which is not so great.

In this commit I've replaced that plugin with Rollup. Rollup is
capable of rolling imported modules into a single UMD closure,
so only "crushinator" itself is exposed in the global object.